requirements.txt
streamlit 
pillow
google-generativeai
langchain
langchain-openai
langchain-groq
requests

